Replace prometheus on the appliance with @/metrics
Summary:
This adds a basic bgmetrics module, which replaces prometheus.  Rather
than spawning a web server in each daemon to feed the collected stats to
a non-existent prometheus daemon, we now periodically send updated stats
to @/metrics/<daemon_name>/<stat>.  This reduces the footprint of each
binary and makes the collected stats available in the cloud as well as
on the appliance.

Test Plan:
currently running on nils-wrt, with stats being read on
svc0.b10e.net

Differential Revision: https://ph0.b10e.net/D904

+++ golang/src/bg/ap.watchd/sampler.go
@@ -1,5 +1,5 @@
 /*
- * COPYRIGHT 2018 Brightgate Inc.  All rights reserved.
+ * COPYRIGHT 2020 Brightgate Inc.  All rights reserved.
  *
  * This copyright notice is Copyright Management Information under 17 USC 1202
  * and is included to protect this work and deter copyright infringement.
@@ -392,7 +392,7 @@ func sampleLoop(state *samplerState) {
 					state.iface, s.PacketsDropped,
 					s.PacketsReceived)
 				lastDropped = s.PacketsDropped
-				metrics.missedPkts.Add(float64(delta))
+				metrics.missedPkts.Add(int64(delta))
 			}
 		}
 	}
